Fix: unittest compile warning

+++ tests/config_test.cc
@@ -42,15 +42,15 @@ TEST(Namespace, Add) {
   config.requirepass = "foobared";
   std::vector<std::string> namespaces= {"n1", "n2", "n3", "n4"};
   std::vector<std::string> tokens = {"t1", "t2", "t3", "t4"};
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     EXPECT_TRUE(config.AddNamespace(namespaces[i], tokens[i]).IsOK());
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     std::string token;
     config.GetNamespace(namespaces[i], &token);
     EXPECT_EQ(token, tokens[i]);
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     auto s = config.AddNamespace(namespaces[i], tokens[i]);
     EXPECT_FALSE(s.IsOK());
     EXPECT_EQ(s.Msg(), "the token has already exists");
@@ -66,23 +66,23 @@ TEST(Namespace, Set) {
   std::vector<std::string> namespaces= {"n1", "n2", "n3", "n4"};
   std::vector<std::string> tokens = {"t1", "t2", "t3", "t4"};
   std::vector<std::string> new_tokens = {"nt1", "nt2'", "nt3", "nt4"};
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     auto s = config.SetNamespace(namespaces[i], tokens[i]);
     EXPECT_FALSE(s.IsOK());
     EXPECT_EQ(s.Msg(), "the namespace was not found");
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     EXPECT_TRUE(config.AddNamespace(namespaces[i], tokens[i]).IsOK());
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     std::string token;
     config.GetNamespace(namespaces[i], &token);
     EXPECT_EQ(token, tokens[i]);
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     EXPECT_TRUE(config.SetNamespace(namespaces[i], new_tokens[i]).IsOK());
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     std::string token;
     config.GetNamespace(namespaces[i], &token);
     EXPECT_EQ(token, new_tokens[i]);
@@ -94,10 +94,10 @@ TEST(Namespace, Delete) {
   config.requirepass = "foobared";
   std::vector<std::string> namespaces= {"n1", "n2", "n3", "n4"};
   std::vector<std::string> tokens = {"t1", "t2", "t3", "t4"};
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     EXPECT_TRUE(config.AddNamespace(namespaces[i], tokens[i]).IsOK());
   }
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     std::string token;
     config.GetNamespace(namespaces[i], &token);
     EXPECT_EQ(token, tokens[i]);
@@ -119,14 +119,14 @@ TEST(Namespace, RewriteNamespaces) {
   config.Load(path) ;
   std::vector<std::string> namespaces= {"n1", "n2", "n3", "n4"};
   std::vector<std::string> tokens = {"t1", "t2", "t3", "t4"};
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     EXPECT_TRUE(config.AddNamespace(namespaces[i], tokens[i]).IsOK());
   }
   EXPECT_TRUE(config.Rewrite().IsOK());
   Config new_config;
   auto s = new_config.Load(path) ;
   std::cout << s.Msg() << std::endl;
-  for(int i = 0; i < namespaces.size(); i++) {
+  for(size_t i = 0; i < namespaces.size(); i++) {
     std::string token;
     new_config.GetNamespace(namespaces[i], &token);
     EXPECT_EQ(token, tokens[i]);
